+++ src/Functions/Functions/ChangeLog.md
@@ -23,7 +23,7 @@
 * [Upgraded code generator](https://go.microsoft.com/fwlink/?linkid=2340249)
 * Validated SKU in `Update-AzFunctionApp`
 * Fixed cloud portability using dynamic `StorageEndpointSuffix` [#29034]
-* Fixed `Update-AzFunctionApp`
+* Fixed handling of 'unknown' runtime to return a formatted name when `FUNCTIONS_WORKER_RUNTIME` is missing or empty [#29630]
 
 ## Version 4.3.2
 * Preannounced breaking changes. Please refer to https://go.microsoft.com/fwlink/?linkid=2333229